// packages/revenuecat-api/src/generated/customerResources.gen.ts
export interface paths {
// ...
  "/projects/{project_id}/customers/{customer_id}/purchases": {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /**
     * Get a list of purchases associated with a customer
     * @description This endpoint requires the following permission(s): <code>customer_information:purchases:read</code>. This endpoint belongs to the <strong>Customer Information</strong> domain, which has a default rate limit of <strong>480 requests per minute</strong>.
     */
    get: operations["list-purchases"];
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
  "/projects/{project_id}/customers/{customer_id}/subscriptions": {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /**
     * Get a list of subscriptions associated with a customer
     * @description This endpoint requires the following permission(s): <code>customer_information:subscriptions:read</code>. This endpoint belongs to the <strong>Customer Information</strong> domain, which has a default rate limit of <strong>480 requests per minute</strong>.
     */
    get: operations["list-subscriptions"];
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
// ...
}
export type webhooks = Record<string, never>;
export interface components {
  schemas: {
    /**
     * CustomerActiveEntitlementList
     * @description List of the entitlements currently active for the customer. This property is only available in the "Get a customer" endpoint.
     */
    ListCustomerActiveEntitlements: {
      /**
       * @description String representing the object's type. Objects of the same type share the same value. Always has the value `list`.
       * @enum {string}
       */
      object: "list";
      /** @description Details about each object. */
      items: components["schemas"]["CustomerEntitlement"][];
// ...
    };
// ...
  };
  responses: {
    /** @description Bad request */
    BadRequest: {
      headers: {
        [name: string]: unknown;
      };
      content: {
        "application/json": components["schemas"]["Error"] & {
          /** @enum {string} */
          type?: "parameter_error" | "invalid_request";
        };
      };
    };
    /** @description Unauthorized */
    Unauthorized: {
      headers: {
        [name: string]: unknown;
      };
      content: {
        "application/json": components["schemas"]["Error"] & {
          /** @enum {string} */
          type?: "authentication_error";
        };
      };
    };
    /** @description Access denied */
    Forbidden: {
      headers: {
        [name: string]: unknown;
      };
      content: {
        "application/json": components["schemas"]["Error"] & {
          /** @enum {string} */
          type?: "authorization_error";
        };
      };
    };
    /** @description Not found */
    NotFound: {
      headers: {
        [name: string]: unknown;
      };
      content: {
        "application/json": components["schemas"]["Error"] & {
          /** @enum {string} */
          type?: "resource_missing";
        };
      };
    };
    /** @description Locked */
    Locked: {
      headers: {
        [name: string]: unknown;
      };
      content: {
        "application/json": components["schemas"]["Error"] & {
          /** @enum {string} */
          type?: "resource_locked_error";
        };
      };
    };
    /** @description The request could not be completed because the rate limiting domain for this endpoint is currently at its limit for this project. */
    RateLimited: {
      headers: {
        "RevenueCat-Rate-Limit-Current-Usage": components["headers"]["RateLimitCurrentUsage"];
        "RevenueCat-Rate-Limit-Current-Limit": components["headers"]["RateLimitCurrentLimit"];
        "Retry-After": components["headers"]["RateLimitRetryAfter"];
        [name: string]: unknown;
      };
      content: {
        "application/json": components["schemas"]["Error"] & {
          /** @enum {string} */
          type?: "rate_limit_error";
        };
      };
    };
    /** @description Internal server error */
    InternalError: {
      headers: {
        [name: string]: unknown;
      };
      content: {
        "application/json": components["schemas"]["Error"] & {
          /** @enum {string} */
          type?: "server_error";
        };
      };
    };
    /** @description Conflict */
    Conflict: {
      headers: {
        [name: string]: unknown;
      };
      content: {
        "application/json": components["schemas"]["Error"] & {
          /** @enum {string} */
          type?: "resource_already_exists" | "idempotency_error" | "invalid_request";
        };
      };
    };
    /** @description Unprocessable entity */
    UnprocessableEntity: {
      headers: {
        [name: string]: unknown;
      };
      content: {
        "application/json": components["schemas"]["Error"] & {
          /** @enum {string} */
          type?:
            | "unprocessable_entity_error"
            | "parameter_error"
            | "store_error"
            | "entity_references_archived_entities";
        };
      };
    };
  };
// ...
}
export type $defs = Record<string, never>;
export interface operations {
  "list-customer-active-entitlements": {
// ...
    responses: {
      /** @description Success */
      200: {
        headers: {
          "RevenueCat-Rate-Limit-Current-Usage": components["headers"]["RateLimitCurrentUsage"];
          "RevenueCat-Rate-Limit-Current-Limit": components["headers"]["RateLimitCurrentLimit"];
          [name: string]: unknown;
        };
        content: {
          "application/json": components["schemas"]["ListCustomerActiveEntitlements"];
        };
      };
      400: components["responses"]["BadRequest"];
      401: components["responses"]["Unauthorized"];
      403: components["responses"]["Forbidden"];
      404: components["responses"]["NotFound"];
      423: components["responses"]["Locked"];
      429: components["responses"]["RateLimited"];
      500: components["responses"]["InternalError"];
      503: components["responses"]["InternalError"];
    };
  };
// ...
}